requirements.txt
beautifulsoup4==4.4.1
cloudinary==1.1.3
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-autoslug==1.9.3
django-mptt==0.7.4
django-mptt-admin==0.3.0
gunicorn==19.3.0
lxml==3.4.4
psycopg2==2.6.1
requests==2.8.1
schedule==0.3.2
six==1.10.0
static3==0.6.1
wheel==0.24.0
